+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### [Unreleased][unreleased]
 
+### [0.4.0] - 2018-08-21
 #### Changed
 - Dropped support for PHP 5.6 and 7.0 [BC break]
 
@@ -43,7 +44,8 @@
 #### Added
 - Wrapper for \Redis
 
-[unreleased]: https://github.com/lulco/redis-proxy/compare/0.3.1...HEAD
+[unreleased]: https://github.com/lulco/redis-proxy/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/lulco/redis-proxy/compare/0.3.1...0.4.0
 [0.3.1]: https://github.com/lulco/redis-proxy/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/lulco/redis-proxy/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/lulco/redis-proxy/compare/0.2.0...0.2.1